Separate EC item creation, along with the constants
